Modularization and DRYing of LLM provider Processors.
